+++ damus/Models/UserSettingsStore.swift
@@ -148,7 +148,7 @@ class UserSettingsStore: ObservableObject {
 
     @Published var dm_notification: Bool {
         didSet {
-            UserDefaults.standard.set(repost_notification, forKey: "dm_notification")
+            UserDefaults.standard.set(dm_notification, forKey: "dm_notification")
         }
     }
 
@@ -160,13 +160,25 @@ class UserSettingsStore: ObservableObject {
 
     @Published var notification_only_from_following: Bool {
         didSet {
-            UserDefaults.standard.set(like_notification, forKey: "notification_only_from_following")
+            UserDefaults.standard.set(notification_only_from_following, forKey: "notification_only_from_following")
+        }
+    }
+    
+    @Published var translate_dms: Bool {
+        didSet {
+            UserDefaults.standard.set(translate_dms, forKey: "translate_dms")
         }
     }
 
     @Published var truncate_timeline_text: Bool {
         didSet {
             UserDefaults.standard.set(truncate_timeline_text, forKey: "truncate_timeline_text")
+        }
+    }
+    
+    @Published var truncate_mention_text: Bool {
+        didSet {
+            UserDefaults.standard.set(truncate_mention_text, forKey: "truncate_mention_text")
         }
     }
 
@@ -269,7 +281,9 @@ class UserSettingsStore: ObservableObject {
         like_notification = UserDefaults.standard.object(forKey: "like_notification") as? Bool ?? true
         dm_notification = UserDefaults.standard.object(forKey: "dm_notification") as? Bool ?? true
         notification_only_from_following = UserDefaults.standard.object(forKey: "notification_only_from_following") as? Bool ?? false
+        translate_dms = UserDefaults.standard.object(forKey: "translate_dms") as? Bool ?? false
         truncate_timeline_text = UserDefaults.standard.object(forKey: "truncate_timeline_text") as? Bool ?? false
+        truncate_mention_text = UserDefaults.standard.object(forKey: "truncate_mention_text") as? Bool ?? false
         disable_animation = should_disable_image_animation()
         auto_translate = UserDefaults.standard.object(forKey: "auto_translate") as? Bool ?? true
         show_only_preferred_languages = UserDefaults.standard.object(forKey: "show_only_preferred_languages") as? Bool ?? false