+++ CHANGELOG.md
@@ -1,3 +1,54 @@
+## [1.4.1-6] - 2023-04-06
+
+### Added
+
+- Custom hashtags for #bitcoin, #nostr and #coffeechain (William Casarin)
+
+### Changed
+
+- Disable translations in DMs by default (William Casarin)
+
+### Fixed
+
+- Don't show Translating... if we're not actually translating (William Casarin)
+
+
+[1.4.1-6]: https://github.com/damus-io/damus/releases/tag/v1.4.1-6
+
+## [1.4.1-4] - 2023-04-06
+
+### Added
+
+- Cache translations (William Casarin)
+
+### Fixed
+
+- Fix translation text popping (William Casarin)
+- Fix broken auto-translations (William Casarin)
+- Fix extraneous padding on some image posts (William Casarin)
+- Fix crash in relay list view (William Casarin)
+
+[1.4.1-4]: https://github.com/damus-io/damus/releases/tag/v1.4.1-4
+
+## [1.4.1-3] - 2023-04-05
+
+### Added
+
+- Added text truncation settings (William Casarin)
+
+### Changed
+
+- Rename block to mute (William Casarin)
+
+### Fixed
+
+- Reduce chopping of images (mainvolume)
+- Fix some notification settings not saving (William Casarin)
+- Fix broken camera uploads (again) (Joel Klabo)
+
+
+[1.4.1-3]: https://github.com/damus-io/damus/releases/tag/v1.4.1-3
+
 ## [1.4.1-2] - 2023-04-04
 
 ### Added
@@ -902,4 +953,5 @@
 
 
 [0.1.2]: https://github.com/damus-io/damus/releases/tag/v0.1.2
+